+++ src/Rask.Server/RaskEndpointExtensions.cs
@@ -199,6 +199,7 @@ public static class RaskEndpointExtensions
         services.AddScoped<IMediaQuery, MediaQuery>();
         services.AddScoped<ISpeechSynthesis, SpeechSynthesis>();
         services.AddScoped<IScreenInfo, ScreenInfoReader>();
+        services.AddScoped<IStorageEstimator, StorageEstimator>();
         services.AddScoped<ICookies, Cookies>();
         services.AddScoped<IPermissions, Permissions>();
         services.AddScoped<IVibration, Vibration>();